mile-knowledge/projects/wj/task_controller/requirements.txt
adbutils==1.2.9
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aioprocessing==2.0.1
aiosignal==1.3.1
altgraph==0.17.3
annotated-types==0.7.0
ansicon==1.89.0
anyio==4.5.2
apkutils2==1.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asn1==2.8.0
asttokens==3.0.0
async-timeout==5.0.1
attrdict==2.0.1
attrs==25.1.0
blessed==1.21.0
bpylist2==4.1.1
cached-property==1.5.2
capstone==5.0.6
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
cigam==0.0.3
click==8.1.8
colorama==0.4.6
colored==1.4.4
coloredlogs==15.0.1
concurrent-log-handler==0.9.22
configparser==7.2.0
construct==2.10.70
cryptography==43.0.3
daemonize==2.5.0
decorator==5.1.1
Deprecated==1.2.13
deprecation==2.1.0
developer_disk_image==0.2.0
dnspython==2.7.0
EasyProcess==1.1
editor==1.6.6
email_validator==2.2.0
entrypoint2==1.1
enum-compat==0.0.3
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.2.0
facebook-wda==1.5.1
fastapi==0.115.8
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
filelock==3.11.0
frozenlist==1.5.0
future==1.0.0
gevent==22.10.2
gpxpy==1.5.0
greenlet==2.0.2
h11==0.14.0
h2==3.2.0
hexdump==3.3
hjson==3.1.0
hpack==3.0.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
humanfriendly==10.0
humanize==4.10.0
hyperframe==5.2.0
idna==2.10
ifaddr==0.2.0
imageio==2.35.1
inquirer3==0.6.1
ipsw-parser==1.4.3
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jinxed==1.3.0
keystone-engine==0.9.2
la-panic==0.5.0
loguru==0.7.0
logzero==1.7.0
lxml==4.9.2
lzfse==0.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
minio==7.2.10
mss==9.0.2
multidict==6.1.0
nest-asyncio==1.6.0
numpy==1.24.2
opack==0.1.0
opack2==0.0.1
openpyxl==3.1.2
orjson==3.11.1
packaging==20.9
pandas==2.0.0
parameter-decorators==0.0.2
parso==0.8.4
pefile==2023.2.7
Pillow==9.5.0
plumbum==1.9.0
portalocker==2.7.0
progress==1.6
prompt_toolkit==3.0.51
propcache==0.2.0
psutil==5.9.4
pure-python-adb==0.3.0.dev0
pure_eval==0.2.3
py==1.11.0
pyasn1==0.5.0
pycparser==2.21
pycrashreport==1.2.6
pycryptodome==3.21.0
pydantic==2.10.6
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.27.2
pyelftools==0.29
PyGetWindow==0.0.9
Pygments==2.19.2
pygnuutils==0.1.1
pyimg4==0.8.8
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
pykdebugparser==1.2.7
pylzss==0.3.4
pymobiledevice3==4.22.1
pymongo==3.4.0
pyOpenSSL==24.2.1
pyparsing==3.0.9
pyreadline3==3.5.4
PyRect==0.2.0
pyscreenshot==3.1
python-dateutil==2.8.2
python-dotenv==1.1.1
python-gitlab==6.2.0
python-multipart==0.0.20
python-pcapng==2.1.1
pytun-pmd3==2.2.2
pytz==2023.3
pyusb==1.3.1
pywin32==306
pywin32-ctypes==0.2.0
PyYAML==6.0.2
qh3==1.5.3
readchar==4.2.1
remotezip==0.12.3
remotezip2==0.0.2
requests==2.32.4
requests-toolbelt==1.0.0
retry==0.9.2
retry2==0.9.5
rfc3986==1.5.0
rich==14.1.0
rich-toolkit==0.14.9
rignore==0.6.4
rpyc==6.0.2
runs==1.2.2
sentry-sdk==2.34.0
shellingham==1.5.4
simple-tornado==0.2.2
simplejson==3.19.1
six==1.16.0
sniffio==1.3.1
srptools==1.0.1
sslpsk-pmd3==1.0.3
stack-data==0.6.3
starlette==0.44.0
tabulate==0.9.0
termcolor==3.1.0
tidevice==0.12.10
tornado==6.3
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2023.3
uiautomator2==2.16.23
ujson==5.10.0
urllib3==1.26.15
uvicorn==0.33.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
weditor==0.6.4
whichcraft==0.6.1
win32-setctime==1.1.0
WMI==1.5.1
wrapt==1.15.0
wsproto==1.2.0
xmltodict==0.13.0
xmod==1.8.1
xonsh==0.19.9
yarl==1.15.2
zeroconf==0.147.0
zope.event==4.6
zope.interface==6.0
